+++ docs/changelog.md
@@ -4,7 +4,7 @@ Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<pa
 
 ## HardPy 0.10.0
 
-* Add the `[stand_cloud]` section to the **harpy.toml** configuration file.
+* Add the `[stand_cloud]` section to the **hardpy.toml** configuration file.
 * Add the `StandCloudLoader` class to append the test result to the **StandCloud**.
 * Add support for **StandCloud** login and logout with `sc-login` and `sc-logout` commands.
 * Add **alert** field in **statestore** database.